requirements.txt
astroid==2.3.3
attrs==19.3.0
aws-cdk.assets==1.25.0
aws-cdk.aws-apigateway==1.25.0
aws-cdk.aws-applicationautoscaling==1.25.0
aws-cdk.aws-autoscaling==1.25.0
aws-cdk.aws-autoscaling-common==1.25.0
aws-cdk.aws-autoscaling-hooktargets==1.25.0
aws-cdk.aws-certificatemanager==1.25.0
aws-cdk.aws-cloudformation==1.25.0
aws-cdk.aws-cloudfront==1.25.0
aws-cdk.aws-cloudwatch==1.25.0
aws-cdk.aws-cognito==1.25.0
aws-cdk.aws-dynamodb==1.25.0
aws-cdk.aws-dynamodb-global==1.25.0
aws-cdk.aws-ec2==1.25.0
aws-cdk.aws-ecr==1.25.0
aws-cdk.aws-ecr-assets==1.25.0
aws-cdk.aws-ecs==1.25.0
aws-cdk.aws-elasticloadbalancing==1.25.0
aws-cdk.aws-elasticloadbalancingv2==1.25.0
aws-cdk.aws-events==1.25.0
aws-cdk.aws-iam==1.25.0
aws-cdk.aws-kms==1.25.0
aws-cdk.aws-lambda==1.25.0
aws-cdk.aws-logs==1.25.0
aws-cdk.aws-route53==1.25.0
aws-cdk.aws-route53-targets==1.25.0
aws-cdk.aws-s3==1.25.0
aws-cdk.aws-s3-assets==1.25.0
aws-cdk.aws-s3-deployment==1.25.0
aws-cdk.aws-sam==1.25.0
aws-cdk.aws-secretsmanager==1.25.0
aws-cdk.aws-servicediscovery==1.25.0
aws-cdk.aws-sns==1.25.0
aws-cdk.aws-sns-subscriptions==1.25.0
aws-cdk.aws-sqs==1.25.0
aws-cdk.aws-ssm==1.25.0
aws-cdk.aws-stepfunctions==1.25.0
aws-cdk.aws-stepfunctions-tasks==1.25.0
aws-cdk.core==1.25.0
aws-cdk.custom-resources==1.25.0
aws-cdk.cx-api==1.25.0
aws-cdk.region-info==1.25.0
cattrs==1.0.0
colorama==0.4.3
importlib-resources==1.0.2
isort==4.3.21
jsii==0.22.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
publication==0.0.3
pylint==2.4.4
python-dateutil==2.8.1
six==1.14.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
wrapt==1.11.2